Refactoring: Move classes from bisq.core.encoding.canonical to bisq.common.encoding.canonical
Move the schema-based canonical encoder to common so Proto.serializeForHash can use canonical bytes for any model that explicitly implements Canonical without creating a p2p dependency on core.
