+++ common/src/main/java/bisq/common/encoding/canonical/CanonicalMapEntryIterator.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.encoding.canonical;
+package bisq.common.encoding.canonical;
 
 import java.util.Iterator;
 import java.util.List;
