+++ common/src/main/java/bisq/common/encoding/canonical/CanonicalWriter.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.encoding.canonical;
+package bisq.common.encoding.canonical;
 
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
