+++ common/src/main/java/bisq/common/encoding/canonical/TreeMapIterator.java
@@ -15,7 +15,7 @@
  * License along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.encoding.canonical;
+package bisq.common.encoding.canonical;
 
 import java.util.Comparator;
 import java.util.Iterator;
