+++ core/src/main/java/bisq/core/dao/state/model/blockchain/BaseBlock.java
@@ -18,7 +18,7 @@
 package bisq.core.dao.state.model.blockchain;
 
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.CanonicalSchema;
 
 import java.util.Optional;
