+++ core/src/main/java/bisq/core/dao/state/model/blockchain/BaseTxOutput.java
@@ -17,8 +17,8 @@
 
 package bisq.core.dao.state.model.blockchain;
 
-import bisq.core.encoding.canonical.Canonical;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.Canonical;
+import bisq.common.encoding.canonical.CanonicalSchema;
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
 
 import bisq.common.util.JsonExclude;
