+++ core/src/main/java/bisq/core/dao/state/model/blockchain/TxOutput.java
@@ -18,9 +18,9 @@
 package bisq.core.dao.state.model.blockchain;
 
 import bisq.core.dao.node.parser.TempTxOutput;
-import bisq.core.encoding.canonical.Canonical;
-import bisq.core.encoding.canonical.CanonicalEncoder;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.Canonical;
+import bisq.common.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalSchema;
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
 
 import bisq.common.proto.persistable.PersistablePayload;
