+++ core/src/main/java/bisq/core/dao/state/model/blockchain/TxOutputType.java
@@ -17,7 +17,7 @@
 
 package bisq.core.dao.state.model.blockchain;
 
-import bisq.core.encoding.canonical.CanonicalEnum;
+import bisq.common.encoding.canonical.CanonicalEnum;
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
 
 import bisq.common.proto.ProtoUtil;
