+++ core/src/main/java/bisq/core/dao/state/model/governance/CompensationProposal.java
@@ -22,9 +22,9 @@ import bisq.core.dao.governance.proposal.IssuanceProposal;
 import bisq.core.dao.governance.proposal.ProposalType;
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
 import bisq.core.dao.state.model.blockchain.TxType;
-import bisq.core.encoding.canonical.CanonicalEncoder;
-import bisq.core.encoding.canonical.CanonicalSchema;
-import bisq.core.encoding.canonical.TreeMapIterator;
+import bisq.common.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.TreeMapIterator;
 
 import bisq.common.app.Version;
 import bisq.common.util.CollectionUtils;
@@ -32,7 +32,6 @@ import bisq.common.util.CollectionUtils;
 import org.bitcoinj.core.Coin;
 
 import java.util.Date;
-import java.util.Map;
 import java.util.Optional;
 import java.util.TreeMap;
