+++ core/src/main/java/bisq/core/dao/state/model/governance/DecryptedBallotsWithMerits.java
@@ -20,9 +20,9 @@ package bisq.core.dao.state.model.governance;
 import bisq.core.dao.governance.merit.MeritConsensus;
 import bisq.core.dao.state.DaoStateService;
 import bisq.core.dao.state.model.ImmutableDaoStateModel;
-import bisq.core.encoding.canonical.Canonical;
-import bisq.core.encoding.canonical.CanonicalEncoder;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.Canonical;
+import bisq.common.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalSchema;
 
 import bisq.common.proto.persistable.PersistablePayload;
 import bisq.common.util.Utilities;
