+++ core/src/test/java/bisq/core/dao/node/parser/TempTxOutputCanonicalEncoderTest.java
@@ -20,8 +20,8 @@ package bisq.core.dao.node.parser;
 import bisq.core.dao.node.full.RawTxOutput;
 import bisq.core.dao.state.model.blockchain.PubKeyScript;
 import bisq.core.dao.state.model.blockchain.TxOutputType;
-import bisq.core.encoding.canonical.CanonicalEncoder;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalSchema;
 
 import com.google.protobuf.ByteString;
