+++ core/src/test/java/bisq/core/dao/state/model/blockchain/BlockchainCanonicalEncoderTest.java
@@ -17,7 +17,7 @@
 
 package bisq.core.dao.state.model.blockchain;
 
-import bisq.core.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalEncoder;
 
 import com.google.protobuf.ByteString;
