+++ core/src/test/java/bisq/core/dao/state/model/blockchain/PubKeyScriptCanonicalEncoderTest.java
@@ -17,8 +17,8 @@
 
 package bisq.core.dao.state.model.blockchain;
 
-import bisq.core.encoding.canonical.CanonicalEncoder;
-import bisq.core.encoding.canonical.CanonicalSchema;
+import bisq.common.encoding.canonical.CanonicalEncoder;
+import bisq.common.encoding.canonical.CanonicalSchema;
 
 import org.junit.jupiter.api.Test;
